Breakdown linode types duck.
